requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
autopep8==1.7.0
backcall==0.2.0
boto3==1.24.58
botocore==1.27.58
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
decorator==5.1.1
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.23.1
executing==0.8.3
freezegun==1.2.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
inflection==0.5.1
ipython==8.4.0
jedi==0.18.1
jmespath==1.0.1
jsonschema==4.9.1
matplotlib-inline==0.1.3
outcome==1.2.0
parso==0.8.3
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pybrowsers==0.5.0
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stack-data==0.3.0
toml==0.10.2
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.10
wcwidth==0.2.5
webdriver-manager==3.8.0
wsproto==1.1.0
